django/requirements/base.txt
Django==1.6.1
Embedly==0.5.0
Pillow==2.2.2
South==0.8.4
boto==2.21.2
django-celery==3.1.1
django-crispy-forms==1.4.0
django-recaptcha==0.0.6
django-storages==1.1.8
psycopg2==2.5.1
redis==2.8.0
tldextract==1.3.1
